pyproject.toml
[project]
name = "llm-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.3.16",
    "langchain-openai>=0.3.3",
    "openai>=1.61.0",
]
